github-pipelines/benchmark-upload/benchmark-upload-requirements.txt
PyYAML==6.0.1
pymongo==4.7.3